fix: review findings — query optimization, auth hardening, cents conversion
C2: eliminated duplicate trait/option queries in public menu endpoint
H4: D1 .batch() for single round-trip queries in listVisible/_hydrate
M1: rate limiting on login challenge/verify endpoints
M3: sanitized WebAuthn error messages (no internal errors to client)
M7: Zod schemas and types updated for integer cents
M9: PRAGMA foreign_keys uses exec() instead of prepare().run()
L3: scheduled cleanup handler for expired sessions/challenges
L13: removed unused PaginatedResponse type

+++ backend/src/validation/schemas.ts
@@ -58,7 +58,7 @@ export const languageCreateSchema = z.object({
 
 export const menuItemUpdateSchema = z.object({
   is_visible: z.boolean().optional(),
-  base_price: z.number().min(0).optional(),
+  base_price: z.number().int().min(0).optional(), // integer cents
 });
 
 export const reorderSchema = z.object({
@@ -94,7 +94,7 @@ export const optionCreateSchema = z.object({
 
 // Update option price
 export const optionUpdateSchema = z.object({
-  price_delta: z.number().optional(),
+  price_delta: z.number().int().optional(), // integer cents
 });
 
 // Upsert a setting value
